Add `KernelArg::to_array`
