+++ tests/test_utils.h
@@ -5,6 +5,25 @@
 
 #include "kernel_launcher/kernel.h"
 
+struct point3 {
+    unsigned long long x;
+    unsigned long long y;
+    unsigned long long z;
+
+    bool operator==(point3 v) const {
+        return x == v.x && y == v.y && z == v.z;
+    }
+};
+
+namespace kernel_launcher {
+template<>
+struct IntoKernelArg<point3> {
+    static KernelArg convert(point3 value) {
+        return KernelArg::for_scalar<point3>(value);
+    }
+};
+}  // namespace kernel_launcher
+
 inline std::string assets_directory() {
     std::string assets_dir = __FILE__;
     assets_dir = assets_dir.substr(0, assets_dir.rfind('/'));
